feat: クライアントコンポーネントをサーバーコンポーネントに移行
ローディングエフェクト排除・APIルート廃止・Supabaseクエリのサーバー集約・RLS有効化のための変更。

変更内容:
- games/page.tsx, games/[id]/page.tsx, games/new/page.tsx をサーバーコンポーネントに変換
- games/[id]/edit/page.tsx をサーバーコンポーネントに変換（認証をサーバー側で確認）
- players/page.tsx をサーバーコンポーネントに変換（PlayersClient に CRUD UI を分離）
- stats/page.tsx をサーバーコンポーネントに変換（StatsClient にタブ/ソート UI を分離）
- games/[id]/actions.ts, players/actions.ts: Server Actions でミューテーションを実装
- games/[id]/delete-button.tsx: 削除ボタンを小さなクライアントコンポーネントに分離
- game-editor.tsx: onBack 未指定時にデフォルトナビゲーションを追加
- 不要な API ルートを削除（auth/status, stats, players GET/POST, games GET/POST, players/[id], games/[id] DELETE）
- scripts/003_enable_rls_all_tables.sql: 全テーブルの RLS ポリシーを追加
