+++ app/[teamId]/players/players-client.tsx
@@ -1,0 +1,190 @@
+"use client"
+
+import { useState } from "react"
+import { PlusCircle, Users, Loader2, X, Pencil, Trash2, Check } from "lucide-react"
+import { sortPlayersByNumber } from "@/lib/sort-utils"
+import { addPlayer, updatePlayer, deletePlayer } from "./actions"
+
+interface Player {
+  id: string
+  name: string
+  number?: string
+}
+
+interface PlayersClientProps {
+  initialPlayers: Player[]
+  isAdmin: boolean
+  teamId: string
+}
+
+export function PlayersClient({ initialPlayers, isAdmin, teamId }: PlayersClientProps) {
+  const [players, setPlayers] = useState<Player[]>(initialPlayers)
+  const [isAdding, setIsAdding] = useState(false)
+  const [newName, setNewName] = useState("")
+  const [newNumber, setNewNumber] = useState("")
+  const [isSaving, setIsSaving] = useState(false)
+  const [editingId, setEditingId] = useState<string | null>(null)
+  const [editName, setEditName] = useState("")
+  const [editNumber, setEditNumber] = useState("")
+
+  const handleAdd = async () => {
+    if (!newName.trim()) {
+      alert("名前を入力してください")
+      return
+    }
+    setIsSaving(true)
+    try {
+      const player = await addPlayer(teamId, newName.trim(), newNumber.trim() || null)
+      setPlayers((prev) => sortPlayersByNumber([...prev, player]))
+      setNewName("")
+      setNewNumber("")
+      setIsAdding(false)
+    } catch (error) {
+      alert(error instanceof Error ? error.message : "保存に失敗しました")
+    } finally {
+      setIsSaving(false)
+    }
+  }
+
+  const handleEdit = (player: Player) => {
+    setEditingId(player.id)
+    setEditName(player.name)
+    setEditNumber(player.number?.toString() || "")
+  }
+
+  const handleSaveEdit = async () => {
+    if (!editingId || !editName.trim()) return
+    setIsSaving(true)
+    try {
+      const updated = await updatePlayer(teamId, editingId, editName.trim(), editNumber.trim() || null)
+      setPlayers((prev) => sortPlayersByNumber(prev.map((p) => (p.id === editingId ? updated : p))))
+      setEditingId(null)
+    } catch (error) {
+      alert(error instanceof Error ? error.message : "更新に失敗しました")
+    } finally {
+      setIsSaving(false)
+    }
+  }
+
+  const handleDelete = async (id: string) => {
+    if (!confirm("この選手を削除しますか？")) return
+    try {
+      await deletePlayer(teamId, id)
+      setPlayers((prev) => prev.filter((p) => p.id !== id))
+    } catch (error) {
+      alert(error instanceof Error ? error.message : "削除に失敗しました")
+    }
+  }
+
+  return (
+    <main className="min-h-screen bg-gradient-to-b from-slate-100 to-slate-200">
+      <div className="mx-auto max-w-6xl p-4 md:p-6">
+        <h1 className="mb-6 text-xl font-bold text-slate-800">選手一覧</h1>
+
+        {isAdmin && !isAdding && (
+          <button
+            onClick={() => setIsAdding(true)}
+            className="mb-6 flex w-full items-center justify-center gap-2 rounded-xl bg-blue-600 px-4 py-3 text-white shadow-md transition-all hover:bg-blue-700 active:scale-[0.98]"
+          >
+            <PlusCircle className="h-5 w-5" />
+            <span className="font-bold">選手を追加</span>
+          </button>
+        )}
+
+        {isAdding && (
+          <div className="mb-6 rounded-xl bg-white p-4 shadow-md">
+            <div className="mb-4 flex items-center justify-between">
+              <h2 className="font-bold text-slate-800">新しい選手を追加</h2>
+              <button onClick={() => setIsAdding(false)} className="rounded-lg p-1 text-slate-400 hover:bg-slate-100 hover:text-slate-600">
+                <X className="h-5 w-5" />
+              </button>
+            </div>
+            <div className="flex flex-wrap gap-3">
+              <input
+                type="text"
+                value={newName}
+                onChange={(e) => setNewName(e.target.value)}
+                placeholder="名前"
+                className="flex-1 rounded-lg border border-slate-300 px-3 py-2 text-sm focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500"
+              />
+              <input
+                type="text"
+                value={newNumber}
+                onChange={(e) => setNewNumber(e.target.value)}
+                placeholder="背番号"
+                maxLength={3}
+                className="w-24 rounded-lg border border-slate-300 px-3 py-2 text-sm focus:border-blue-500 focus:outline-none focus:ring-1 focus:ring-blue-500"
+              />
+              <button
+                onClick={handleAdd}
+                disabled={isSaving}
+                className="flex items-center gap-2 rounded-lg bg-blue-600 px-4 py-2 text-sm font-bold text-white transition-all hover:bg-blue-700 disabled:opacity-50"
+              >
+                {isSaving ? <Loader2 className="h-4 w-4 animate-spin" /> : "追加"}
+              </button>
+            </div>
+          </div>
+        )}
+
+        {players.length === 0 ? (
+          <div className="rounded-2xl bg-white p-8 text-center shadow-md">
+            <Users className="mx-auto mb-3 h-12 w-12 text-slate-300" />
+            <p className="text-slate-500">まだ選手が登録されていません</p>
+          </div>
+        ) : (
+          <div className="rounded-2xl bg-white shadow-md">
+            <div className="divide-y divide-slate-100">
+              {players.map((player) => (
+                <div key={player.id} className="flex items-center gap-4 px-4 py-3">
+                  {editingId === player.id ? (
+                    <>
+                      <input
+                        type="text"
+                        value={editNumber}
+                        onChange={(e) => setEditNumber(e.target.value)}
+                        placeholder="番号"
+                        maxLength={3}
+                        className="w-16 rounded-lg border border-slate-300 px-2 py-1 text-sm text-center"
+                      />
+                      <input
+                        type="text"
+                        value={editName}
+                        onChange={(e) => setEditName(e.target.value)}
+                        className="flex-1 rounded-lg border border-slate-300 px-3 py-1 text-sm"
+                      />
+                      <button onClick={handleSaveEdit} disabled={isSaving} className="rounded-lg bg-emerald-600 p-2 text-white hover:bg-emerald-700 disabled:opacity-50">
+                        {isSaving ? <Loader2 className="h-4 w-4 animate-spin" /> : <Check className="h-4 w-4" />}
+                      </button>
+                      <button onClick={() => setEditingId(null)} className="rounded-lg bg-slate-200 p-2 text-slate-600 hover:bg-slate-300">
+                        <X className="h-4 w-4" />
+                      </button>
+                    </>
+                  ) : (
+                    <>
+                      {player.number !== undefined && player.number !== null && (
+                        <div className="flex h-10 w-10 items-center justify-center rounded-full bg-slate-100 text-sm font-bold text-slate-600">
+                          {player.number}
+                        </div>
+                      )}
+                      <span className="flex-1 font-medium text-slate-800">{player.name}</span>
+                      {isAdmin && (
+                        <div className="flex gap-2">
+                          <button onClick={() => handleEdit(player)} className="rounded-lg p-2 text-slate-400 hover:bg-slate-100 hover:text-blue-600">
+                            <Pencil className="h-4 w-4" />
+                          </button>
+                          <button onClick={() => handleDelete(player.id)} className="rounded-lg p-2 text-slate-400 hover:bg-slate-100 hover:text-red-600">
+                            <Trash2 className="h-4 w-4" />
+                          </button>
+                        </div>
+                      )}
+                    </>
+                  )}
+                </div>
+              ))}
+            </div>
+          </div>
+        )}
+      </div>
+    </main>
+  )
+}
